Move split_values function into contricleaner/lib/helpers
